+++ examples/basic/src/model-function/generate-text-logging-example.ts
@@ -12,7 +12,7 @@ dotenv.config();
     new OpenAITextGenerationModel({
       model: "text-davinci-003",
       temperature: 0.7,
-      maxTokens: 500,
+      maxCompletionTokens: 500,
       observers: [new ConsoleLogger()],
     }),
     "Write a short story about a robot learning to love:\n\n"