+++ examples/basic/src/model-provider/llamacpp/llamacpp-text-generation-example.ts
@@ -2,7 +2,10 @@ import { LlamaCppTextGenerationModel, generateText } from "modelfusion";
 
 (async () => {
   const text = await generateText(
-    new LlamaCppTextGenerationModel({ nPredict: 256, temperature: 0.7 }),
+    new LlamaCppTextGenerationModel({
+      maxCompletionTokens: 256,
+      temperature: 0.7,
+    }),
     "Write a short story about a robot learning to love:\n\n"
   );