+++ examples/basic/src/observability/helicone/helicone-openai-chat-stream-text-example.ts
@@ -18,7 +18,7 @@ if (HELICONE_API_KEY == null) {
       },
       model: "gpt-3.5-turbo",
       temperature: 0.7,
-      maxTokens: 500,
+      maxCompletionTokens: 500,
     }),
     [
       OpenAIChatMessage.system(