+++ src/composed-function/summarize/summarizeRecursivelyWithTextGenerationAndTokenSplitting.ts
@@ -21,7 +21,7 @@ export async function summarizeRecursivelyWithTextGenerationAndTokenSplitting<
     model,
     prompt,
     tokenLimit = model.contextWindowSize -
-      (model.maxCompletionTokens ?? model.contextWindowSize / 4),
+      (model.settings.maxCompletionTokens ?? model.contextWindowSize / 4),
     join,
   }: {
     text: string;